feat. støtte double for stillingsprosent og antallDagerPerUke

+++ src/main/java/no/nav/tag/tiltaksgjennomforingprosess/domene/avtale/Avtale.java
@@ -39,8 +39,8 @@ public class Avtale {
     private String veilederTlf;
     private String oppfolging;
     private String tilrettelegging;
-    private Integer stillingprosent;
-    private Integer antallDagerPerUke;
+    private Double stillingprosent;
+    private Double antallDagerPerUke;
     private String enhetOppfolging;
     private Integer versjon;
     private String mentorFornavn;
@@ -117,5 +117,3 @@ public class Avtale {
         return this.versjon > 1;
     }
 }
-
-
